# ddtrace/tracer: improve payload performance