This change improves the payload performance on high throughput by
reducing the number of allocations. The old implementation creates new
slices on each call, while the new implementation reuses the same slice
every time.

     name                      old time/op    new time/op    delta
     PayloadThroughput/10K-8     12.8µs ± 0%    12.6µs ± 0%   -0.94%
     PayloadThroughput/100K-8     134µs ± 0%      95µs ± 0%  -28.90%
     PayloadThroughput/1MB-8      701µs ± 0%     582µs ± 0%  -17.05%

     name                      old speed      new speed      delta
     PayloadThroughput/10K-8    812MB/s ± 0%   819MB/s ± 0%   +0.95%
     PayloadThroughput/100K-8   774MB/s ± 0%  1088MB/s ± 0%  +40.65%
     PayloadThroughput/1MB-8   1.48GB/s ± 0%  1.78GB/s ± 0%  +20.56%

     name                      old alloc/op   new alloc/op   delta
     PayloadThroughput/10K-8     28.5kB ± 0%    28.5kB ± 0%   -0.06%
     PayloadThroughput/100K-8     278kB ± 0%     278kB ± 0%   -0.01%
     PayloadThroughput/1MB-8     2.85MB ± 0%    2.85MB ± 0%   -0.00%

     name                      old allocs/op  new allocs/op  delta
     PayloadThroughput/10K-8       3.00 ± 0%      1.00 ± 0%  -66.67%
     PayloadThroughput/100K-8      3.00 ± 0%      1.00 ± 0%  -66.67%
     PayloadThroughput/1MB-8       3.00 ± 0%      1.00 ± 0%  -66.67%

This change also improve tests to cover all msgpack array sizes.